refactor: レビュー指摘対応（メニュー定義の共通化・TopPageへ改名）
- ナビゲーションのメニュー定義（MenuItem / menuItems / getVisibleMenuItems）を
  sidebar コンポーネントから nav-menu/menu-items.ts へ切り出し、sidebar への
  逆依存（および型のみの循環参照）を解消
- TrendDiaryTopPage を TopPage に改名
