+++ application/apps/web/src/client/components/ui/navigation/nav-menu/index.tsx
@@ -1,7 +1,7 @@
 import { SheetClose } from '@/client/components/shadcn/sheet'
 import { SidebarMenu, SidebarMenuButton, SidebarMenuItem } from '@/client/components/shadcn/sidebar'
-import type { MenuItem } from '@/client/components/ui/layout/sidebar'
 import { AnchorLink } from '@/client/components/ui/navigation/link'
+import type { MenuItem } from './menu-items'
 
 interface NavMenuProps {
   variant: 'sidebar' | 'sheet'
